Fix frontend production calling backend api
