+++ frontend/package.json
@@ -8,7 +8,6 @@
     "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
-  "homepage": ".",
   "eslintConfig": {
     "extends": "react-app"
   },
