+++ frontend/src/utils/api-handlers/apiurl.ts
@@ -1,0 +1,1 @@
+export default process.env.REACT_APP_API;
